refac(TimestampFactory):
REFACTOR: deduplicate implementations of TimestampFactory and IdFactory.

Also remove line in cfg.py to prevent circular imports.
All done. Here's a summary of the changes:
1. gen_epix/commondb/domain/enum.py — Added TimestampFactory and IdFactory with required imports (datetime, uuid, ulid)
2. gen_epix/seqdb/domain/enum.py — Removed duplicate definitions; replaced with re-exports from commondb/domain/enum
3. gen_epix/omopdb/domain/enum.py — Same as seqdb
4. gen_epix/commondb/config/factory.py — Deleted (only contained these two classes)
5. gen_epix/commondb/config/init.py — Updated import source to commondb/domain/enum
6. gen_epix/commondb/config/cfg.py — Updated import source to commondb/domain/enum The seqdb and omopdb enums still re-export TimestampFactory and IdFactory (with as aliases) to maintain backward compatibility with any wildcard-import consumers, matching the pattern already used for RoleSet.
