+++ gen_epix/commondb/config/__init__.py
@@ -2,5 +2,3 @@
 from gen_epix.commondb.config.cfg import AppCfg as AppCfg
 from gen_epix.commondb.config.cfg import BaseAppCfg as BaseAppCfg
 from gen_epix.commondb.config.cfg import SettingsManager as SettingsManager
-from gen_epix.commondb.config.factory import IdFactory as IdFactory
-from gen_epix.commondb.config.factory import TimestampFactory as TimestampFactory
