--- gen_epix/commondb/config/factory.py
@@ -1,14 +1,0 @@
-import datetime
-import uuid
-from enum import Enum
-
-import ulid
-
-
-class TimestampFactory(Enum):
-    DATETIME_NOW = lambda: datetime.datetime.now(datetime.timezone.utc)
-
-
-class IdFactory(Enum):
-    UUID4 = uuid.uuid4
-    ULID = lambda: ulid.api.new().uuid
