+++ gen_epix/commondb/domain/enum.py
@@ -1,4 +1,17 @@
+import datetime
+import uuid
 from enum import Enum, IntEnum
+
+import ulid
+
+
+class TimestampFactory(Enum):
+    DATETIME_NOW = lambda: datetime.datetime.now(datetime.timezone.utc)
+
+
+class IdFactory(Enum):
+    UUID4 = uuid.uuid4
+    ULID = lambda: ulid.api.new().uuid
 
 
 class Role(Enum):
